Moved random_multimodal to _mmvec.py
